+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - `Formal\AccessLayer\Query\Select::limit()`
+- `Formal\AccessLayer\Query\Select::orderBy()`
+- `Formal\AccessLayer\Query\Select\Direction`
 
 ## 2.2.0 - 2023-07-30